mrouted: Refactor, relocate logging funs to log.c
