+++ src/log.c
@@ -1,0 +1,215 @@
+/*
+ * The mrouted program is covered by the license in the accompanying file
+ * named "LICENSE".  Use of the mrouted program represents acceptance of
+ * the terms and conditions listed in that file.
+ *
+ * The mrouted program is COPYRIGHT 1989 by The Board of Trustees of
+ * Leland Stanford Junior University.
+ */
+
+#define SYSLOG_NAMES
+#include "defs.h"
+#include <stdarg.h>
+
+#define LOG_MAX_MSGS	20	/* if > 20/minute then shut up for a while */
+#define LOG_SHUT_UP	600	/* shut up for 10 minutes */
+
+int loglevel = LOG_NOTICE;
+
+static int log_nmsgs = 0;
+
+
+int log_str2lvl(char *level)
+{
+    int i;
+
+    for (i = 0; prioritynames[i].c_name; i++) {
+	size_t len = MIN(strlen(prioritynames[i].c_name), strlen(level));
+
+	if (!strncasecmp(prioritynames[i].c_name, level, len))
+	    return prioritynames[i].c_val;
+    }
+
+    return atoi(level);
+}
+
+void resetlogging(void *arg)
+{
+    int nxttime = 60;
+    void *narg = NULL;
+
+    if (arg == NULL && log_nmsgs > LOG_MAX_MSGS) {
+	nxttime = LOG_SHUT_UP;
+	narg = (void *)&log_nmsgs;	/* just need some valid void * */
+	syslog(LOG_WARNING, "logging too fast, shutting up for %d minutes",
+			LOG_SHUT_UP / 60);
+    } else {
+	log_nmsgs = 0;
+    }
+
+    timer_setTimer(nxttime, resetlogging, narg);
+}
+
+
+
+/*
+ * Open connection to syslog daemon and set initial log level
+ */
+void log_init(void)
+{
+#ifdef LOG_DAEMON
+    openlog("mrouted", LOG_PID, LOG_DAEMON);
+    setlogmask(LOG_UPTO(loglevel));
+#else
+    openlog("mrouted", LOG_PID);
+#endif
+}
+
+
+/*
+ * Log errors and other messages to the system log daemon and to stderr,
+ * according to the severity of the message and the current debug level.
+ * For errors of severity LOG_ERR or worse, terminate the program.
+ */
+void logit(int severity, int syserr, const char *format, ...)
+{
+    va_list ap;
+    static char fmt[211] = "warning - ";
+    char *msg;
+    struct timeval now;
+    time_t now_sec;
+    struct tm *thyme;
+#ifdef RINGBUFFER
+    static int ringbufinit = 0;
+#endif
+
+    va_start(ap, format);
+    vsnprintf(&fmt[10], sizeof(fmt) - 10, format, ap);
+    va_end(ap);
+    msg = (severity == LOG_WARNING) ? fmt : &fmt[10];
+
+#ifdef RINGBUFFER
+    if (!ringbufinit) {
+	int i;
+
+	for (i = 0; i < NLOGMSGS; i++) {
+	    logmsg[i] = malloc(LOGMSGSIZE);
+	    if (logmsg[i] == 0) {
+		syslog(LOG_ERR, "Out of memory");
+		exit(1);
+	    }
+	    *logmsg[i] = 0;
+	}
+	ringbufinit = 1;
+    }
+    gettimeofday(&now, NULL);
+    now_sec = now.tv_sec;
+    thyme = localtime(&now_sec);
+    snprintf(logmsg[logmsgno++], LOGMSGSIZE, "%02d:%02d:%02d.%03ld %s err %d",
+	     thyme->tm_hour, thyme->tm_min, thyme->tm_sec,
+	     now.tv_usec / 1000, msg, syserr);
+    logmsgno %= NLOGMSGS;
+    if (severity <= LOG_NOTICE)
+#endif
+    /*
+     * Log to stderr if we haven't forked yet and it's a warning or worse,
+     * or if we're debugging.
+     */
+    if (haveterminal && (debug || severity <= LOG_WARNING)) {
+	gettimeofday(&now, NULL);
+	now_sec = now.tv_sec;
+	thyme = localtime(&now_sec);
+	if (!debug)
+	    fprintf(stderr, "%s: ", PACKAGE_NAME);
+	fprintf(stderr, "%02d:%02d:%02d.%03ld %s", thyme->tm_hour,
+		    thyme->tm_min, thyme->tm_sec, now.tv_usec / 1000, msg);
+	if (syserr == 0)
+	    fprintf(stderr, "\n");
+	else
+	    fprintf(stderr, ": %s\n", strerror(syserr));
+    }
+
+    /*
+     * Always log things that are worse than warnings, no matter what
+     * the log_nmsgs rate limiter says.
+     * Only count things worse than debugging in the rate limiter
+     * (since if you put daemon.debug in syslog.conf you probably
+     * actually want to log the debugging messages so they shouldn't
+     * be rate-limited)
+     */
+    if ((severity < LOG_WARNING) || (log_nmsgs < LOG_MAX_MSGS)) {
+	if (severity < LOG_DEBUG)
+	    log_nmsgs++;
+	if (syserr != 0) {
+	    errno = syserr;
+	    syslog(severity, "%s: %s", msg, strerror(errno));
+	} else
+	    syslog(severity, "%s", msg);
+    }
+
+    if (severity <= LOG_ERR)
+	exit(1);
+}
+
+#ifdef RINGBUFFER
+#define NLOGMSGS 10000
+#define LOGMSGSIZE 200
+char *logmsg[NLOGMSGS];
+static int logmsgno = 0;
+
+void printringbuf(void)
+{
+    FILE *f;
+    int i;
+
+    f = fopen("/var/tmp/mrouted.log", "a");
+    if (f == NULL) {
+	logit(LOG_ERR, errno, "Cannot open /var/tmp/mrouted.log");
+	/*NOTREACHED*/
+    }
+    fprintf(f, "--------------------------------------------\n");
+
+    i = (logmsgno + 1) % NLOGMSGS;
+
+    while (i != logmsgno) {
+	if (*logmsg[i]) {
+	    fprintf(f, "%s\n", logmsg[i]);
+	    *logmsg[i] = '\0';
+	}
+	i = (i + 1) % NLOGMSGS;
+    }
+
+    fclose(f);
+}
+#endif
+
+#ifdef DEBUG_MFC
+void md_log(int what, uint32_t origin, uint32_t mcastgrp)
+{
+    static FILE *f = NULL;
+    struct timeval tv;
+    uint32_t buf[4];
+
+    if (!f) {
+	if ((f = fopen("/tmp/mrouted.clog", "w")) == NULL) {
+	    logit(LOG_ERR, errno, "open /tmp/mrouted.clog");
+	}
+    }
+
+    gettimeofday(&tv, NULL);
+    buf[0] = tv.tv_sec;
+    buf[1] = what;
+    buf[2] = origin;
+    buf[3] = mcastgrp;
+
+    fwrite(buf, sizeof(uint32_t), 4, f);
+}
+#endif
+
+/**
+ * Local Variables:
+ *  indent-tabs-mode: t
+ *  c-file-style: "ellemtel"
+ *  c-basic-offset: 4
+ * End:
+ */
